Move components into components dir
Transition away from the pattern of separatign components and containers
as per the update in dabramov's post on the subject:
https://medium.com/@dan_abramov/smart-and-dumb-components-7ca2f9a7c7d0
