+++ tdrs-frontend/src/App.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import { GovBanner } from '@trussworks/react-uswds'
-import Routes from './Routes'
+import Routes from './components/Routes'
 
 import '@trussworks/react-uswds/lib/uswds.css'
 import '@trussworks/react-uswds/lib/index.css'
