+++ tdrs-frontend/src/components/Routes/Routes.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import { Switch, Route } from 'react-router-dom'
-import Welcome from './containers/Welcome'
+import Welcome from '../Welcome'
 
 const Routes = () => {
   return (
