+++ tdrs-frontend/src/components/Welcome/Welcome.scss
